LayoutTests/editing/spelling/spelling-changed-text.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/util.js"></script>
</head>
<body>
<div id="container">
  <div id="destination" contentEditable></div>
</div>

<script>

description("Spellcheck should not crash after the text has changed and results are served from cache."
    + " To test manually, launch Chromium compiled with Address Sanitizer, enable 'Ask Google for Suggestions', type 'Spell cheher. Is it broken?', delete the words 'Is it broken?', and context-click on the word 'cheher'."
    + " The test suceeds when the browser does not crash and shows suggestions in the context menu.");

initSpellTest("destination", "Spell cheher. Is it broken?", function(textNode) {
    // Select the text "Is it broken?".
    var deleteRange = document.createRange();
    deleteRange.setStart(textNode, 13);
    deleteRange.setEnd(textNode, 27);
    window.getSelection().removeAllRanges();
    window.getSelection().addRange(deleteRange);
    shouldBeEqualToString("window.getSelection().toString()", " Is it broken?");
    // Del key to delete the text "Is it broken?".
    eventSender.keyDown(String.fromCharCode(0x007F), null);

    // Context click to show the context menu.
    var x = destination.offsetParent.offsetLeft + destination.offsetLeft + 50;
    var y = destination.offsetParent.offsetTop + destination.offsetTop + destination.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    contextMenuElements = eventSender.contextClick();
    // Esc key to hide the context menu.
    eventSender.keyDown(String.fromCharCode(0x001B), null);

    document.getElementById("destination").innerHTML = "";
});

</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
